Conflicts:
	SztucznaIntCw/SztucznaIntCw/MainWindow.cs

+++ SztucznaIntCw/SztucznaIntCw/Classes/Interfaces/IMeal.cs
@@ -4,13 +4,13 @@ namespace SztucznaIntCw.Classes.Interfaces
 {
     public interface IMeal
     {
-       
-        int TotalGramsOfProteins { get; set; }
-        int TotalGramsOfFats { get; set; }
-        int TotalGramsOfCarbs { get; set; }
+
+        decimal TotalGramsOfProteins { get; set; }
+        decimal TotalGramsOfFats { get; set; }
+        decimal TotalGramsOfCarbs { get; set; }
 
         List<IProduct> ProductList { get; }
         //List<IProduct> ProductList { get; }
-        Dictionary<IProduct, float> MealProducts { get; }
+        Dictionary<IProduct, decimal> MealProducts { get; }
     }
 }
